+++ db/cardano-sl-db.cabal
@@ -27,6 +27,8 @@ library
                        Pos.DB.Pure
                        Pos.DB.Sum
                        Pos.DB.Rocks
+                       Pos.DB.Rocks.Functions
+                       Pos.DB.Rocks.Types
 
                        Pos.DB.Txp
                        Pos.DB.Txp.Logic
